--- src/Server/Game/Models/IChangeCache.cs
@@ -1,9 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Imlight.Server.Game.Models;
-
-public interface IChangeCache
-{
-    void EnqueueChange(object change);
-    Task FlushChangesAsync();
-}
