+++ src/Server/Game/Models/WizardSchool.cs
@@ -1,3 +1,8 @@
+/* Copyright (C) Revive101 Development Team - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential.
+ */
+
 namespace Imlight.Server.Game.Models;
 
 public enum WizardSchool
