+++ src/Server/Game/ResultHandlerAttribute.cs
@@ -1,7 +1,11 @@
-using System;
-using WizUnraveler.Cache;
+/* Copyright (C) Revive101 Development Team - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential.
+ */
 
-namespace Imlight.Server.Game.Models;
+using System;
+
+namespace Imlight.Server.Game;
 
 [AttributeUsage(AttributeTargets.Method, AllowMultiple = false)]
 public class ResultHandlerAttribute : Attribute
