+++ src/Server/WizardData/ElementChangeCacheManager.cs
@@ -1,9 +1,15 @@
+/* Copyright (C) Revive101 Development Team - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential.
+ */
+
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Imlight.Server.Game.Models;
 using Raven.Client.Documents;
 
-namespace Imlight.Server.Game.Models;
+namespace Imlight.Server.WizardData;
 
 public class ElementChangeCacheManager
 {
